requirements.txt
# Core dependencies
python-telegram-bot==20.7
geopy==2.4.1
python-dotenv==1.0.0

# Optional: Better date/time handling
pytz==2024.1
